api/src/requirements.txt
flask
flask-restful
flask-SQLAlchemy
psycopg2
pytest
passlib==1.7.1
geoalchemy2==0.6.3
gunicorn==19.9.0
pyjwt
